+++ lib/src/feature_discovery.dart
@@ -32,7 +32,9 @@ class FeatureDiscovery extends StatefulWidget {
 }
 
 class _FeatureDiscoveryState extends State<FeatureDiscovery> {
-  static _FeatureDiscoveryState of(BuildContext context) => context.ancestorStateOfType(TypeMatcher<_FeatureDiscoveryState>()) as _FeatureDiscoveryState;
+  static _FeatureDiscoveryState of(BuildContext context) =>
+      context.ancestorStateOfType(TypeMatcher<_FeatureDiscoveryState>())
+          as _FeatureDiscoveryState;
 
   List<String> steps;
   int activeStepIndex;
@@ -88,79 +90,93 @@ class _FeatureDiscoveryState extends State<FeatureDiscovery> {
 class DescribedFeatureOverlay extends StatefulWidget {
   /// This id must be unique among all the [DescribedFeatureOverlay]s widgets.
   final String featureId;
-  final IconData icon;
-
-  @Deprecated("Replaced by backgroundColor")
-  final Color color;
 
   /// The color of the outside layout, behind texts.
-  /// This parameter replaces the [color] parameter: you should only specify this one.
   /// If null, defaults to [ThemeData.primaryColor].
   final Color backgroundColor;
 
-  /// If null, defaults to current [IconTheme]
-  final Color iconColor;
-
-  /// Color of the target, that is the small circle behind the icon.
+  /// Color of the target, that is the small circle behind the tap target.
   final Color targetColor;
 
-  /// Color for title and text
+  /// Color for title and text.
   final Color textColor;
+
   final String title;
   final String description;
 
-  /// Called when the target is pressed.
-  final Function(VoidCallback onActionCompleted) doAction;
+  /// This is usually an [Icon].
+  /// The final tap target will already have a tap listener to finish each step.
+  ///
+  /// If you want to hit the tap target in integration tests, you should pass a [Key]
+  /// to this [Widget] instead of as the [Key] of [DescribedFeatureOverlay].
+  final Widget tapTarget;
 
-  /// Called just before the FeatureOverlay is displayed.
-  /// The function parameter is actually the callback that triggers the display of the overlay.
-  /// If not null, the callback MUST be called in order for the overlay to be displayed.
-  final Function(VoidCallback onActionCompleted) prepareAction;
   final Widget child;
   final ContentOrientation contentLocation;
   final bool enablePulsingAnimation;
 
-  /// Function to execute when the overlay is dismissed (when the user taps outside of it).
-  /// If not null, the callback MUST be called in order for the overlay to be dismissed.
-  final Function(VoidCallback onActionCompleted) onDismissAction;
+  /// Called just before the overlay is displayed.
+  /// This function needs to return a [bool], either from an `async` scope
+  /// or as a [Future].
+  ///
+  /// If this function returns `false`, this step will be marked complete
+  /// and therefore be skipped, i.e. it will not be opened.
+  /// In this case, we try to open the next step.
+  ///
+  /// When the [Future] finishes and evaluates to `true`, this step will be shown.
+  final Future<bool> Function() onOpen;
+
+  /// Called whenever the user taps outside the overlay area.
+  /// This function needs to return a [bool], either from an `async` scope
+  /// or as a [Future].
+  ///
+  /// If the function returns `false`, nothing happens. If it returns `true`,
+  /// all of the current steps are dismissed.
+  final Future<bool> Function() onDismiss;
+
+  /// Called when the tap target is tapped.
+  /// Whenever the [Future] this function returns is finished, the feature discovery
+  /// will continue and the next step will try to open after a closing animation.
+  final Future<void> Function() onTargetTap;
 
   const DescribedFeatureOverlay({
     Key key,
     @required this.featureId,
-    @required this.icon,
-    this.color,
+    @required this.tapTarget,
     this.backgroundColor,
-    this.iconColor,
     this.targetColor = Colors.white,
     this.textColor = Colors.white,
     this.title,
     this.description,
     @required this.child,
-    this.doAction,
-    this.prepareAction,
+    this.onTargetTap,
+    this.onDismiss,
+    this.onOpen,
     this.contentLocation = ContentOrientation.trivial,
     this.enablePulsingAnimation = true,
-    this.onDismissAction,
   })  : assert(featureId != null),
-        assert(icon != null),
+        assert(tapTarget != null),
         assert(child != null),
         assert(contentLocation != null),
         assert(enablePulsingAnimation != null),
         assert(targetColor != null),
         assert(textColor != null),
-        assert(color == null || backgroundColor == null, "[color] parameter has been replaced by [backgroundColor]: as they're the same, you should only specify one"),
         super(key: key);
 
   @override
-  _DescribedFeatureOverlayState createState() => _DescribedFeatureOverlayState();
+  _DescribedFeatureOverlayState createState() =>
+      _DescribedFeatureOverlayState();
 }
 
-class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with TickerProviderStateMixin {
+class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay>
+    with TickerProviderStateMixin {
   Size screenSize;
-  double statusBarHeight;
-  bool showOverlay = false;
-  _OverlayState state = _OverlayState.closed;
-  double transitionPercent = 1.0;
+
+  bool showOverlay;
+
+  _OverlayState state;
+
+  double transitionProgress;
 
   AnimationController openController;
   AnimationController activationController;
@@ -169,6 +185,12 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
 
   @override
   void initState() {
+    showOverlay = false;
+
+    state = _OverlayState.closed;
+
+    transitionProgress = 1;
+
     initAnimationControllers();
     super.initState();
   }
@@ -183,29 +205,37 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
   }
 
   void initAnimationControllers() {
-    openController = AnimationController(vsync: this, duration: Duration(milliseconds: 250))
-      ..addListener(() => setState(() => transitionPercent = openController.value))
-      ..addStatusListener(
-        (AnimationStatus status) {
-          if (status == AnimationStatus.forward)
-            setState(() => state = _OverlayState.opening);
-          else if (status == AnimationStatus.completed) pulseController?.forward(from: 0.0);
-        },
-      );
+    openController =
+        AnimationController(vsync: this, duration: Duration(milliseconds: 250))
+          ..addListener(
+              () => setState(() => transitionProgress = openController.value))
+          ..addStatusListener(
+            (AnimationStatus status) {
+              if (status == AnimationStatus.forward)
+                setState(() => state = _OverlayState.opening);
+              else if (status == AnimationStatus.completed)
+                pulseController?.forward(from: 0.0);
+            },
+          );
 
     if (widget.enablePulsingAnimation) {
-      pulseController = AnimationController(vsync: this, duration: Duration(milliseconds: 1000))
-        ..addListener(() => setState(() => transitionPercent = pulseController.value))
+      pulseController = AnimationController(
+          vsync: this, duration: Duration(milliseconds: 1000))
+        ..addListener(
+            () => setState(() => transitionProgress = pulseController.value))
         ..addStatusListener(
           (AnimationStatus status) {
             if (status == AnimationStatus.forward)
               setState(() => state = _OverlayState.pulsing);
-            else if (status == AnimationStatus.completed) pulseController.forward(from: 0.0);
+            else if (status == AnimationStatus.completed)
+              pulseController.forward(from: 0.0);
           },
         );
     }
-    activationController = AnimationController(vsync: this, duration: Duration(milliseconds: 250))
-      ..addListener(() => setState(() => transitionPercent = activationController.value))
+    activationController = AnimationController(
+        vsync: this, duration: Duration(milliseconds: 250))
+      ..addListener(
+          () => setState(() => transitionProgress = activationController.value))
       ..addStatusListener(
         (AnimationStatus status) {
           switch (status) {
@@ -213,11 +243,7 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
               setState(() => state = _OverlayState.activating);
               break;
             case AnimationStatus.completed:
-              void Function() callback = () => FeatureDiscovery.markStepComplete(context, widget.featureId);
-              if (widget.doAction == null)
-                callback();
-              else
-                widget.doAction(callback);
+              FeatureDiscovery.markStepComplete(context, widget.featureId);
               break;
             default:
               break;
@@ -225,19 +251,16 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
         },
       );
 
-    dismissController = AnimationController(vsync: this, duration: Duration(milliseconds: 250))
-      ..addListener(() => setState(() => transitionPercent = dismissController.value))
+    dismissController = AnimationController(
+        vsync: this, duration: Duration(milliseconds: 250))
+      ..addListener(
+          () => setState(() => transitionProgress = dismissController.value))
       ..addStatusListener(
         (AnimationStatus status) {
           if (status == AnimationStatus.forward)
             setState(() => state = _OverlayState.dismissing);
-          else if (status == AnimationStatus.completed) {
-            void Function() callback = () => FeatureDiscovery.dismiss(context);
-            if (widget.onDismissAction == null)
-              callback();
-            else
-              widget.onDismissAction(callback);
-          }
+          else if (status == AnimationStatus.completed)
+            FeatureDiscovery.dismiss(context);
         },
       );
   }
@@ -246,14 +269,14 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
   void didChangeDependencies() {
     super.didChangeDependencies();
     screenSize = MediaQuery.of(context).size;
-    statusBarHeight = MediaQuery.of(context).viewInsets.top;
+
     showOverlayIfActiveStep();
   }
 
   void show() {
     final String activeStep = FeatureDiscovery.activeStep(context);
 
-    // The activeStep might have changed by now (if prepareAction called the callback later).
+    // The activeStep might have changed by now because onOpen is asynchronous.
     if (activeStep != widget.featureId) return;
 
     openController.forward(from: 0.0);
@@ -276,12 +299,18 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
       // which is already showing the overlay for this feature.
       if (_FeatureDiscoveryState.of(context).stepShown) return;
 
-      // This needs to be set here and not in show to prevent multiple prepareAction
+      // This needs to be set here and not in show to prevent multiple onOpen
       // calls to stack up (if there are multiple widgets with the same feature id).
       _FeatureDiscoveryState.of(context).stepShown = true;
 
-      if (widget.prepareAction != null)
-        widget.prepareAction(show);
+      if (widget.onOpen != null)
+        widget.onOpen().then((shouldOpen) {
+          assert(shouldOpen != null);
+          if (shouldOpen)
+            show();
+          else // Mark this step complete to go to the next step as this step is not shown.
+            FeatureDiscovery.markStepComplete(context, widget.featureId);
+        });
       else
         show();
       return;
@@ -294,12 +323,18 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
     }
   }
 
-  void activate() {
+  void activate() async {
+    if (widget.onTargetTap != null) await widget.onTargetTap();
     pulseController?.stop();
     activationController.forward(from: 0.0);
   }
 
-  void dismiss() {
+  void dismiss() async {
+    if (widget.onDismiss != null) {
+      final shouldDismiss = await widget.onDismiss();
+      assert(shouldDismiss != null);
+      if (!shouldDismiss) return;
+    }
     pulseController?.stop();
     dismissController.forward(from: 0.0);
   }
@@ -317,15 +352,15 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
         ),
         _Background(
           state: state,
-          transitionPercent: transitionPercent,
+          transitionProgress: transitionProgress,
           anchor: anchor,
-          color: (widget.backgroundColor ?? widget.color) ?? Theme.of(context).primaryColor,
+          color: widget.backgroundColor ?? Theme.of(context).primaryColor,
           screenSize: screenSize,
           orientation: widget.contentLocation,
         ),
         _Content(
           state: state,
-          transitionPercent: transitionPercent,
+          transitionProgress: transitionProgress,
           anchor: anchor,
           screenSize: screenSize,
           // this parameter is not used
@@ -340,18 +375,17 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
         ),
         _Pulse(
           state: state,
-          transitionPercent: transitionPercent,
+          transitionProgress: transitionProgress,
           anchor: anchor,
           color: widget.targetColor,
         ),
-        _TouchTarget(
+        _TapTarget(
           state: state,
-          transitionPercent: transitionPercent,
+          transitionProgress: transitionProgress,
           anchor: anchor,
-          icon: widget.icon,
-          iconColor: widget.iconColor,
-          backgroundColor: widget.targetColor,
+          color: widget.targetColor,
           onPressed: activate,
+          child: widget.tapTarget,
         ),
       ],
     );
@@ -371,7 +405,7 @@ class _DescribedFeatureOverlayState extends State<DescribedFeatureOverlay> with 
 
 class _Background extends StatelessWidget {
   final _OverlayState state;
-  final double transitionPercent;
+  final double transitionProgress;
   final Offset anchor;
   final Color color;
   final Size screenSize;
@@ -383,13 +417,13 @@ class _Background extends StatelessWidget {
     @required this.color,
     @required this.screenSize,
     @required this.state,
-    @required this.transitionPercent,
+    @required this.transitionProgress,
     @required this.orientation,
   })  : assert(anchor != null),
         assert(color != null),
         assert(screenSize != null),
         assert(state != null),
-        assert(transitionPercent != null),
+        assert(transitionProgress != null),
         assert(orientation != null),
         super(key: key);
 
@@ -407,15 +441,19 @@ class _Background extends StatelessWidget {
 
   double radius() {
     final bool isBackgroundCentered = isCloseToTopOrBottom(anchor);
-    final double backgroundRadius = Math.min(screenSize.width, screenSize.height) * (isBackgroundCentered ? 1.0 : 0.7);
+    final double backgroundRadius =
+        Math.min(screenSize.width, screenSize.height) *
+            (isBackgroundCentered ? 1.0 : 0.7);
     switch (state) {
       case _OverlayState.opening:
-        final double adjustedPercent = const Interval(0.0, 0.8, curve: Curves.easeOut).transform(transitionPercent);
+        final double adjustedPercent =
+            const Interval(0.0, 0.8, curve: Curves.easeOut)
+                .transform(transitionProgress);
         return backgroundRadius * adjustedPercent;
       case _OverlayState.activating:
-        return backgroundRadius + transitionPercent * 40.0;
+        return backgroundRadius + transitionProgress * 40.0;
       case _OverlayState.dismissing:
-        return backgroundRadius * (1 - transitionPercent);
+        return backgroundRadius * (1 - transitionProgress);
       default:
         return backgroundRadius;
     }
@@ -433,25 +471,37 @@ class _Background extends StatelessWidget {
       Offset endingBackgroundPosition;
       switch (orientation) {
         case ContentOrientation.trivial:
-          endingBackgroundPosition =
-              Offset(width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0), anchor.dy + (isOnTopHalfOfScreen(anchor) ? -(width / 2.0) + 40.0 : (width / 2.0) - 40.0));
+          endingBackgroundPosition = Offset(
+              width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0),
+              anchor.dy +
+                  (isOnTopHalfOfScreen(anchor)
+                      ? -(width / 2.0) + 40.0
+                      : (width / 2.0) - 40.0));
           break;
         case ContentOrientation.above:
-          endingBackgroundPosition = Offset(width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0), anchor.dy - (width / 2.0) + 40.0);
+          endingBackgroundPosition = Offset(
+              width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0),
+              anchor.dy - (width / 2.0) + 40.0);
           break;
         case ContentOrientation.below:
-          endingBackgroundPosition = Offset(width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0), anchor.dy + (width / 2.0) - 40.0);
+          endingBackgroundPosition = Offset(
+              width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0),
+              anchor.dy + (width / 2.0) - 40.0);
           break;
       }
 
       switch (state) {
         case _OverlayState.opening:
-          final double adjustedPercent = const Interval(0.0, 0.8, curve: Curves.easeOut).transform(transitionPercent);
-          return Offset.lerp(startingBackgroundPosition, endingBackgroundPosition, adjustedPercent);
+          final double adjustedPercent =
+              const Interval(0.0, 0.8, curve: Curves.easeOut)
+                  .transform(transitionProgress);
+          return Offset.lerp(startingBackgroundPosition,
+              endingBackgroundPosition, adjustedPercent);
         case _OverlayState.activating:
           return endingBackgroundPosition;
         case _OverlayState.dismissing:
-          return Offset.lerp(endingBackgroundPosition, startingBackgroundPosition, transitionPercent);
+          return Offset.lerp(endingBackgroundPosition,
+              startingBackgroundPosition, transitionProgress);
         default:
           return endingBackgroundPosition;
       }
@@ -461,15 +511,21 @@ class _Background extends StatelessWidget {
   double backgroundOpacity() {
     switch (state) {
       case _OverlayState.opening:
-        final double adjustedPercent = const Interval(0.0, 0.3, curve: Curves.easeOut).transform(transitionPercent);
+        final double adjustedPercent =
+            const Interval(0.0, 0.3, curve: Curves.easeOut)
+                .transform(transitionProgress);
         return 0.96 * adjustedPercent;
 
       case _OverlayState.activating:
-        final double adjustedPercent = const Interval(0.1, 0.6, curve: Curves.easeOut).transform(transitionPercent);
+        final double adjustedPercent =
+            const Interval(0.1, 0.6, curve: Curves.easeOut)
+                .transform(transitionProgress);
 
         return 0.96 * (1 - adjustedPercent);
       case _OverlayState.dismissing:
-        final double adjustedPercent = const Interval(0.2, 1.0, curve: Curves.easeOut).transform(transitionPercent);
+        final double adjustedPercent =
+            const Interval(0.2, 1.0, curve: Curves.easeOut)
+                .transform(transitionProgress);
         return 0.96 * (1 - adjustedPercent);
       default:
         return 0.96;
@@ -487,7 +543,9 @@ class _Background extends StatelessWidget {
       child: Container(
         width: 2 * radius(),
         height: 2 * radius(),
-        decoration: BoxDecoration(shape: BoxShape.circle, color: color.withOpacity(backgroundOpacity())),
+        decoration: BoxDecoration(
+            shape: BoxShape.circle,
+            color: color.withOpacity(backgroundOpacity())),
       ),
     );
   }
@@ -495,36 +553,35 @@ class _Background extends StatelessWidget {
 
 class _Pulse extends StatelessWidget {
   final _OverlayState state;
-  final double transitionPercent;
+  final double transitionProgress;
   final Offset anchor;
   final Color color;
 
-  const _Pulse({
-    Key key, 
-    @required this.state, 
-    @required this.transitionPercent, 
-    @required this.anchor, 
-    @required this.color
-  }) : 
-    assert(state != null),
-    assert(transitionPercent != null),
-    assert(anchor != null),
-    assert(color != null),
-    super(key: key);
+  const _Pulse(
+      {Key key,
+      @required this.state,
+      @required this.transitionProgress,
+      @required this.anchor,
+      @required this.color})
+      : assert(state != null),
+        assert(transitionProgress != null),
+        assert(anchor != null),
+        assert(color != null),
+        super(key: key);
 
   double radius() {
     switch (state) {
       case _OverlayState.pulsing:
         double expandedPercent;
-        if (transitionPercent >= 0.3 && transitionPercent <= 0.8) {
-          expandedPercent = (transitionPercent - 0.3) / 0.5;
+        if (transitionProgress >= 0.3 && transitionProgress <= 0.8) {
+          expandedPercent = (transitionProgress - 0.3) / 0.5;
         } else {
           expandedPercent = 0.0;
         }
         return 44.0 + (35.0 * expandedPercent);
       case _OverlayState.dismissing:
       case _OverlayState.activating:
-        return 0.0; //(44.0 + 35.0) * (1.0 - transitionPercent);
+        return 0.0; //(44.0 + 35.0) * (1.0 - transitionProgress);
       default:
         return 0.0;
     }
@@ -533,11 +590,12 @@ class _Pulse extends StatelessWidget {
   double opacity() {
     switch (state) {
       case _OverlayState.pulsing:
-        final double percentOpaque = 1.0 - ((transitionPercent.clamp(0.3, 0.8) - 0.3) / 0.5);
+        final double percentOpaque =
+            1.0 - ((transitionProgress.clamp(0.3, 0.8) - 0.3) / 0.5);
         return (percentOpaque * 0.75).clamp(0.0, 1.0);
       case _OverlayState.activating:
       case _OverlayState.dismissing:
-        return 0.0; //((1.0 - transitionPercent) * 0.5).clamp(0.0, 1.0);
+        return 0.0; //((1.0 - transitionProgress) * 0.5).clamp(0.0, 1.0);
       default:
         return 0.0;
     }
@@ -546,56 +604,57 @@ class _Pulse extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return state == _OverlayState.closed
-      ? Container(height: 0, width: 0)
-      : CenterAbout(
-          position: anchor,
-          child: Container(
-            width: radius() * 2,
-            height: radius() * 2,
-            decoration: BoxDecoration(
-              shape: BoxShape.circle,
-              color: color.withOpacity(opacity()),
+        ? Container(height: 0, width: 0)
+        : CenterAbout(
+            position: anchor,
+            child: Container(
+              width: radius() * 2,
+              height: radius() * 2,
+              decoration: BoxDecoration(
+                shape: BoxShape.circle,
+                color: color.withOpacity(opacity()),
+              ),
             ),
-          ),
-        );
+          );
   }
 }
 
-class _TouchTarget extends StatelessWidget {
+class _TapTarget extends StatelessWidget {
   final _OverlayState state;
-  final double transitionPercent;
+  final double transitionProgress;
   final Offset anchor;
-  final IconData icon;
-  final Color iconColor;
-  final Color backgroundColor;
+  final Widget child;
+  final Color color;
   final VoidCallback onPressed;
 
-  const _TouchTarget({
+  const _TapTarget({
     Key key,
     @required this.anchor,
-    @required this.icon,
+    @required this.child,
     // The two parameters below can technically be null, so assertions are not made for them,
     // but they are annotated as required to not forget them during development
     // (as this is an internal widget and those parameters should always be specified, even when null)
-    @required this.iconColor,
     @required this.onPressed,
-    @required this.backgroundColor,
+    @required this.color,
     @required this.state,
-    @required this.transitionPercent,
+    @required this.transitionProgress,
   })  : assert(anchor != null),
-        assert(icon != null),
+        assert(child != null),
         assert(state != null),
-        assert(transitionPercent != null),
-        assert(backgroundColor != null),
+        assert(transitionProgress != null),
+        assert(color != null),
         super(key: key);
 
   double opacity() {
     switch (state) {
       case _OverlayState.opening:
-        return const Interval(0.0, 0.3, curve: Curves.easeOut).transform(transitionPercent);
+        return const Interval(0.0, 0.3, curve: Curves.easeOut)
+            .transform(transitionProgress);
       case _OverlayState.activating:
       case _OverlayState.dismissing:
-        return 1.0 - const Interval(0.7, 1.0, curve: Curves.easeOut).transform(transitionPercent);
+        return 1.0 -
+            const Interval(0.7, 1.0, curve: Curves.easeOut)
+                .transform(transitionProgress);
       default:
         return 1.0;
     }
@@ -606,19 +665,19 @@ class _TouchTarget extends StatelessWidget {
       case _OverlayState.closed:
         return 0.0;
       case _OverlayState.opening:
-        return 20.0 + 24.0 * transitionPercent;
+        return 20.0 + 24.0 * transitionProgress;
       case _OverlayState.pulsing:
         double expandedPercent;
-        if (transitionPercent < 0.3)
-          expandedPercent = transitionPercent / 0.3;
-        else if (transitionPercent < 0.6)
-          expandedPercent = 1.0 - ((transitionPercent - 0.3) / 0.3);
+        if (transitionProgress < 0.3)
+          expandedPercent = transitionProgress / 0.3;
+        else if (transitionProgress < 0.6)
+          expandedPercent = 1.0 - ((transitionProgress - 0.3) / 0.3);
         else
           expandedPercent = 0.0;
         return 44.0 + (20.0 * expandedPercent);
       case _OverlayState.activating:
       case _OverlayState.dismissing:
-        return 20.0 + 24.0 * (1 - transitionPercent);
+        return 20.0 + 24.0 * (1 - transitionProgress);
       default:
         return 44.0;
     }
@@ -634,12 +693,9 @@ class _TouchTarget extends StatelessWidget {
         child: Opacity(
           opacity: opacity(),
           child: RawMaterialButton(
-            fillColor: backgroundColor,
+            fillColor: color,
             shape: const CircleBorder(),
-            child: Icon(
-              icon,
-              color: iconColor,
-            ),
+            child: child,
             onPressed: onPressed,
           ),
         ),
@@ -650,7 +706,7 @@ class _TouchTarget extends StatelessWidget {
 
 class _Content extends StatelessWidget {
   final _OverlayState state;
-  final double transitionPercent;
+  final double transitionProgress;
   final Offset anchor;
   final Size screenSize;
   final double touchTargetRadius;
@@ -677,7 +733,7 @@ class _Content extends StatelessWidget {
     @required this.title,
     @required this.description,
     @required this.state,
-    @required this.transitionPercent,
+    @required this.transitionProgress,
     //this.statusBarHeight,
     @required this.orientation,
     @required this.textColor,
@@ -685,7 +741,7 @@ class _Content extends StatelessWidget {
         assert(screenSize != null),
         assert(touchTargetRadius != null),
         assert(state != null),
-        assert(transitionPercent != null),
+        assert(transitionProgress != null),
         assert(orientation != null),
         assert(textColor != null),
         super(key: key);
@@ -704,9 +760,13 @@ class _Content extends StatelessWidget {
 
   DescribedFeatureContentOrientation getContentOrientation(Offset position) {
     if (isCloseToTopOrBottom(position))
-      return isOnTopHalfOfScreen(position) ? DescribedFeatureContentOrientation.below : DescribedFeatureContentOrientation.above;
+      return isOnTopHalfOfScreen(position)
+          ? DescribedFeatureContentOrientation.below
+          : DescribedFeatureContentOrientation.above;
     else
-      return isOnTopHalfOfScreen(position) ? DescribedFeatureContentOrientation.above : DescribedFeatureContentOrientation.below;
+      return isOnTopHalfOfScreen(position)
+          ? DescribedFeatureContentOrientation.above
+          : DescribedFeatureContentOrientation.below;
   }
 
   double opacity() {
@@ -714,11 +774,15 @@ class _Content extends StatelessWidget {
       case _OverlayState.closed:
         return 0.0;
       case _OverlayState.opening:
-        final double adjustedPercent = const Interval(0.6, 1.0, curve: Curves.easeOut).transform(transitionPercent);
+        final double adjustedPercent =
+            const Interval(0.6, 1.0, curve: Curves.easeOut)
+                .transform(transitionProgress);
         return adjustedPercent;
       case _OverlayState.activating:
       case _OverlayState.dismissing:
-        final double adjustedPercent = const Interval(0.0, 0.4, curve: Curves.easeOut).transform(transitionPercent);
+        final double adjustedPercent =
+            const Interval(0.0, 0.4, curve: Curves.easeOut)
+                .transform(transitionProgress);
         return 1.0 - adjustedPercent;
       default:
         return 1.0;
@@ -733,17 +797,25 @@ class _Content extends StatelessWidget {
       return anchor;
     else {
       final Offset startingBackgroundPosition = anchor;
-      final Offset endingBackgroundPosition =
-          Offset(width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0), anchor.dy + (isOnTopHalfOfScreen(anchor) ? -(width / 2) + 40.0 : (width / 20.0) - 40.0));
+      final Offset endingBackgroundPosition = Offset(
+          width / 2.0 + (isOnLeftHalfOfScreen(anchor) ? -20.0 : 20.0),
+          anchor.dy +
+              (isOnTopHalfOfScreen(anchor)
+                  ? -(width / 2) + 40.0
+                  : (width / 20.0) - 40.0));
 
       switch (state) {
         case _OverlayState.opening:
-          final double adjustedPercent = const Interval(0.0, 0.8, curve: Curves.easeOut).transform(transitionPercent);
-          return Offset.lerp(startingBackgroundPosition, endingBackgroundPosition, adjustedPercent);
+          final double adjustedPercent =
+              const Interval(0.0, 0.8, curve: Curves.easeOut)
+                  .transform(transitionProgress);
+          return Offset.lerp(startingBackgroundPosition,
+              endingBackgroundPosition, adjustedPercent);
         case _OverlayState.activating:
           return endingBackgroundPosition;
         case _OverlayState.dismissing:
-          return Offset.lerp(endingBackgroundPosition, startingBackgroundPosition, transitionPercent);
+          return Offset.lerp(endingBackgroundPosition,
+              startingBackgroundPosition, transitionProgress);
         default:
           return endingBackgroundPosition;
       }
@@ -752,12 +824,16 @@ class _Content extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final DescribedFeatureContentOrientation contentOrientation = getContentOrientation(anchor);
+    final DescribedFeatureContentOrientation contentOrientation =
+        getContentOrientation(anchor);
     double contentOffsetMultiplier;
 
     switch (orientation) {
       case ContentOrientation.trivial:
-        contentOffsetMultiplier = contentOrientation == DescribedFeatureContentOrientation.below ? 1.0 : -1.0;
+        contentOffsetMultiplier =
+            contentOrientation == DescribedFeatureContentOrientation.below
+                ? 1.0
+                : -1.0;
         break;
       case ContentOrientation.above:
         contentOffsetMultiplier = -1.0;
@@ -769,9 +845,11 @@ class _Content extends StatelessWidget {
 
     final double width = Math.min(screenSize.width, screenSize.height);
 
-    final double contentY = anchor.dy + contentOffsetMultiplier * (touchTargetRadius + 20);
+    final double contentY =
+        anchor.dy + contentOffsetMultiplier * (touchTargetRadius + 20);
 
-    final double contentFractionalOffset = contentOffsetMultiplier.clamp(-1.0, 0.0);
+    final double contentFractionalOffset =
+        contentOffsetMultiplier.clamp(-1.0, 0.0);
 
     final double dx = centerPosition().dx - width;
     final double contentX = (dx.isNegative) ? 0.0 : dx;
@@ -791,14 +869,23 @@ class _Content extends StatelessWidget {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: <Widget>[
-                    title == null ? const SizedBox(height: 0) : Text(title, style: Theme.of(context).textTheme.title.copyWith(color: textColor)),
+                    title == null
+                        ? const SizedBox(height: 0)
+                        : Text(title,
+                            style: Theme.of(context)
+                                .textTheme
+                                .title
+                                .copyWith(color: textColor)),
                     const SizedBox(height: 8.0),
                     description == null
-                      ? const SizedBox(height: 0)
-                      : Text(
-                          description,
-                          style: Theme.of(context).textTheme.body1.copyWith(color: textColor.withOpacity(0.9)),
-                        ),
+                        ? const SizedBox(height: 0)
+                        : Text(
+                            description,
+                            style: Theme.of(context)
+                                .textTheme
+                                .body1
+                                .copyWith(color: textColor.withOpacity(0.9)),
+                          ),
                   ],
                 ),
               ),
@@ -820,10 +907,13 @@ class _InheritedFeatureDiscovery extends InheritedWidget {
   })  : assert(child != null),
         super(key: key, child: child);
 
-  static _InheritedFeatureDiscovery of(BuildContext context) => context.inheritFromWidgetOfExactType(_InheritedFeatureDiscovery) as _InheritedFeatureDiscovery;
+  static _InheritedFeatureDiscovery of(BuildContext context) =>
+      context.inheritFromWidgetOfExactType(_InheritedFeatureDiscovery)
+          as _InheritedFeatureDiscovery;
 
   @override
-  bool updateShouldNotify(_InheritedFeatureDiscovery old) => old.activeStepId != activeStepId;
+  bool updateShouldNotify(_InheritedFeatureDiscovery old) =>
+      old.activeStepId != activeStepId;
 }
 
 enum DescribedFeatureContentOrientation {